Add (transparent) 2D layers support for menu screens and HUD
You can now see the menu and HUD using the mod! By clearing the main render buffer with a black & transparent color after it has rendered the 3D scene, the layer can capture the HUD that'd normally be rendered on top of the 3D scene later totally separate, and we can use an OpenXR quad projection layer like the ones used for overlays to project the menu and HUD properly into the 3D scene.

I also fixed some of the destructors so that it can exit a bit more gracefully, although its not perfect yet.