Fix stutter, rework some rendering and patches
This finally fixes the stutter that was caused by both the projection offsets and rotation not being properly synchronized.

I also reworked the rendering so that while being theoretically slightly less flexible, it allows for better error checking and also allows for better object destruction that'll allow for proper exiting. The code was also getting pretty messy so this cleans it up slightly.